requirements.txt
torch
numpy<2
pandas
scikit-learn
matplotlib
seaborn
tqdm
pyyaml
kagglehub
